requirements.txt
Flask==3.1.0
flask-cors==5.0.1
gunicorn==21.2.0

# HTTP & API access
requests==2.32.3

# Google Gemini AI
google-generativeai==0.5.2
protobuf==4.25.3

# WHOIS domain lookup
python-whois==0.8.0

# VirusTotal needs base64 + request (already covered)

# SSL parsing & web scraping
beautifulsoup4==4.13.4

# Optional, for parsing date formats from WHOIS if needed
python-dateutil==2.9.0.post0

